Correct avatar image links for server on non-standard port
DVWA is recommended to run in a VM. Accessing VM ports often
involves port forwarding, which may mean that the host will
access non-standard port. In this case, avatar images are not
showing. That is because PHP variable $_SERVER['SERVER_NAME']
does not contain the port.

Use $_SERVER['HTTP_HOST'] instead. This might give additional
bonus in form of new vulnerabilities, as this variable is read
from the client's request.